pyproject.toml
[project]
authors = [
    {name = "Benjamin Dornel", email = "benjamindornel@gmail.com"},
]

requires-python = "<3.14,>=3.10"
dependencies = [
    "streamlit>=1.45.1",
    "pybadges @ git+https://github.com/benjamin-awd/pybadges",
    "plotly>=6.1.2",
    "pdftotext>=3.0.0",
    "pymupdf>=1.26.0",
    "monopoly-core==0.19.6",
    "pymongo>=4.9.0",
    "openai>=1.0.0",
    "python-dotenv>=1.2.2",
]
name = "statement-sensei"
version = "0.10.4"
description = "PDF to CSV conversion for your bank statements"
readme = "README.md"

[project.optional-dependencies]
ocr = [
    "ocrmypdf>=16.10.2",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.isort]
profile = "black"

[tool.pylint]
max-line-length=120
disable = [
    "missing-module-docstring",
    "missing-function-docstring",
    "missing-class-docstring",
    "logging-fstring-interpolation"
]

[tool.pdm.build]
includes = ["webapp"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
build = [
    "git-cliff>=2.8.0",
    "pyinstaller>=6.13.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
fixable = ["ALL"]
exclude = ["tests/*"]

ignore = [
    # modules
    "ANN",

    # specific rules
    "COM812",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D203",
    "D212",
    "E501",
    "ISC001",
    "PD901",
    "PTH123",
    "N999",
    "SLF001"
]
